tests(helpers): move wait functions
